// frontend/apps/allin-ssl/src/components/FlowChart/types.d.ts
export interface FlowNodeProps {
	isEdit: boolean
	type: 'quick' | 'advanced'
	node: FlowNode
	taskComponents?: Record<string, Component>
}

export interface FlowNode {
	id: string
	name: string
	childNode: BaseNodeData
}

// 添加节点选项
export interface NodeSelect {
	title: NodeTitle
	type: NodeNum
	icon: NodeIcon
	selected: boolean
}

// 节点标题配置
export interface NodeTitle {
	name: string
	color?: string
	bgColor?: string
}

// 节点图标配置
export interface NodeIcon {
	name: string
	color?: string
}
// ...
export interface BaseNodeData<T = Record<string, unknown>> {
	type: NodeNum // 节点类型
	id?: string // 节点id，用于编辑
	name: string // 节点名称
	icon?: NodeIcon // 节点图标
	inputs?: Record<string, unknown>[] // 输入，用于连接其他节点的数据
	config?: T // 参数，用于配置当前的节点
	childNode?: BaseNodeData | BranchNodeData | null // 子节点
}

// 分支节点数据
export interface BranchNodeData<T extends Record<string, unknown> = Record<string, unknown>> extends BaseNodeData<T> {
	type: typeof BRANCH // 节点类型
	conditionNodes: ConditionNodeData[] // 子节点
}
// ...
export type NodeNum =
	| typeof START // 开始节点
	| typeof DEFAULT // 默认节点
	| typeof END // 结束节点
	| typeof BRANCH // 分支节点
	| typeof CONDITION // 条件节点
	| typeof EXECUTE_RESULT_BRANCH // 执行结果分支节点（if）
	| typeof EXECUTE_RESULT_CONDITION // 执行结果条件节点
	| typeof UPLOAD // 上传节点（业务）
	| typeof NOTIFY // 通知节点（业务）
	| typeof APPLY // 申请节点（业务）
	| typeof DEPLOY // 部署节点（业务）